Updated collision
