+++ src/entities/Crate.java
@@ -1,4 +1,4 @@
-package Entities;
+package entities;
 
 import gameframework.drawing.Drawable;
 import gameframework.drawing.DrawableImage;
@@ -7,47 +7,48 @@ import gameframework.game.GameData;
 import gameframework.game.GameEntity;
 import gameframework.motion.GameMovable;
 import gameframework.motion.GameMovableDriverDefaultImpl;
+import gameframework.motion.MoveStrategyDefaultImpl;
 import gameframework.motion.MoveStrategyKeyboard;
+import gameframework.motion.blocking.MoveBlocker;
 
 import java.awt.Graphics;
 import java.awt.Rectangle;
 
-public class Player extends GameMovable implements Drawable, GameEntity/*
-																		 * ,
-																		 * Movable
-																		 * ,
-																		 * Overlappable
-																		 */{
+public class Crate extends GameMovable implements Drawable, GameEntity, MoveBlocker{
 
 	protected DrawableImage image;
 	protected GameCanvas canvas;
-
-	public Player(GameData data, int x, int y) {
+	int x;
+	int y;
+	
+	public Crate(GameData data, int x, int y) {
 		super();
-		MoveStrategyKeyboard keyboard = new MoveStrategyKeyboard();
+		
+		MoveStrategyDefaultImpl keyboard = new MoveStrategyDefaultImpl();
 		GameMovableDriverDefaultImpl moveDriver = new GameMovableDriverDefaultImpl();
 		moveDriver.setStrategy(keyboard);
 		moveDriver.setmoveBlockerChecker(data.getMoveBlockerChecker());
 		setDriver(moveDriver);
 
-		position.x = x;
-		position.y = y;
+		this.x = x;
+		this.y = y;
 		this.canvas = data.getCanvas();
-		image = new DrawableImage("/images/Locke.gif", canvas);
-		this.canvas.addKeyListener(keyboard);
+		image = new DrawableImage("/images/Crate.gif", canvas);
 
 	}
 
 	@Override
 	public void draw(Graphics g) {
-		canvas.drawImage(g, image.getImage(), position.x, position.y);
+		canvas.drawImage(g, image.getImage(), x, y);
 		// image.draw(g);
 
 	}
 
 	@Override
 	public Rectangle getBoundingBox() {
-		return new Rectangle(image.getWidth(), 20);
+		Rectangle rectangle = new Rectangle(image.getWidth(), image.getHeight());
+		rectangle.setLocation(x, y);
+		return rectangle;
 	}
 
 	@Override
